week6/more_analysis/test/etl_job/requirements.txt
pymongo
sqlalchemy
psycopg2-binary
pandas
vaderSentiment
sklearn.feature_extraction.text
re
time
